+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - Show download and share button in overdue search screen when search results are loaded
 - Bump itext7 to 7.2.3
 - Fix app crashing when scheduling overdue download worker
+- Handle nullable fields when generating overdue CSV from database
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ### Fixes